Implement Filter Function. Close #1
Implemented a basic "filter" function to filter the concerts and re-render the
viw. The filter currently works by splitting the input string, e.g. `name:
egotronic` becomes ['name', 'egotronic'] where the first one is the property to
search by and the second is the desired value.

Also added mongodb to sample config and updated styles a bit.